Move debug wrappers to App.Fossa.Analyze.Debug
